Separate ItemPopup from ItemPopupContainer
